share_pool: Eliminate compiler warning for atomic64_t in arm32
ascend inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I4EUVI
CVE: NA

-------------------------------------------------

atomic64_t is type long in arm64, and type long long in arm32.
So use %lld to print.

+++ mm/oom_kill.c
@@ -489,9 +489,9 @@ static void dump_tasks(struct mem_cgroup *memcg, const nodemask_t *nodemask)
 			if (!stat)
 				pr_cont("%-9c %-9c ", '-', '-');
 			else {
-				pr_cont("%-9ld %-9ld ", /* byte to KB */
-					atomic64_read(&stat->alloc_size) >> 10,
-					atomic64_read(&stat->k2u_size) >> 10);
+				pr_cont("%-9lld %-9lld ", /* byte to KB */
+					(long long)atomic64_read(&stat->alloc_size) >> 10,
+					(long long)atomic64_read(&stat->k2u_size) >> 10);
 				sp_proc_stat_drop(stat);
 			}
 			pr_cont("%8ld %8lu         %5hd %s\n",
